requirements.txt
Django==1.6.1
South==0.8.4
coverage==3.7.1
django-angular==0.5.0
django-tastypie==0.11.0
execnet==1.1
psycopg2==2.5.2
py==1.4.19
pytest==2.5.1
pytest-cache==1.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.5.2
sqlparse==0.1.10
wsgiref==0.1.2
